New functions, lasts fix before 1.1
2 new functions:
> Client.GetUser, generate a user object from a UserID or use an
existing one
> Server.searchUser, search for a user inside a server with many search
options and plenty of nice informations

All endpoints errors logging have been switched to debug, the client is
responsible to show the errors when it's needed and can't handle it
himself. Less log flooding here.
Fix for a curse side bug, when a message is deleted as soon as he get
received, the curse endpoint can be slow (1/50messages), and not find
it, the lib will retry another time before triggering an error.
UserList has been deleted on Server object because it's impossible to
keep it consistant and way safer to use specific functions to know if
someone is member of the server or not.
Documentation is getting bigger, but still far from done, 1.1 may hit
before pre doc is "complete" and will complete it in the firsts post 1.1
commits.