Message queue for notifier !
It's now possible to chain sending messages without being afraid of
losing these messages, the callback will always now be called with an
error if the message couldn't be sent.
As the backend throttler triggers very easily, it is only possible to
send one message at a time, it will wait for the message confirmation
before sending another one.
If the message is still throttled, the lib will wait 100ms and try to
resend it again.